refactor: lift fix_webview2_child_windows into auroraview-core
Move the WebView2 child-window fix out of NativeBackend into the shared
auroraview-core builder module so both the embedded (Qt/DCC) path and the
standalone top-level path can call it. Switch the per-call tallies from
static atomics to a stack-local Counters struct threaded through the
EnumChildWindows LPARAM, so concurrent calls don't clobber each other.

Add public-export tests covering invalid and arbitrary handles.
