+++ src/webview/backend/native.rs
@@ -429,248 +429,6 @@ impl NativeBackend {
         ProcessResult::Continue
     }
 
-    /// Fix all WebView2 child windows to prevent dragging (Qt6 compatibility)
-    ///
-    /// WebView2 creates multiple child windows (Chrome_WidgetWin_0, Intermediate D3D Window, etc.)
-    /// that may not inherit proper WS_CHILD styles. This function recursively fixes all child
-    /// windows to ensure they cannot be dragged independently.
-    ///
-    /// Additionally, this function subclasses Chrome_WidgetWin_0 and Chrome_WidgetWin_1 windows
-    /// to intercept WM_NCHITTEST messages and force them to return HTCLIENT, preventing any
-    /// drag behavior from the WebView2's internal window handling.
-    ///
-    /// This is especially important for Qt6 where createWindowContainer behavior differs from Qt5.
-    ///
-    /// # Arguments
-    /// * `hwnd` - The top-level WebView window handle
-    #[cfg(target_os = "windows")]
-    pub fn fix_webview2_child_windows(hwnd: isize) {
-        use std::collections::HashMap;
-        use std::ffi::c_void;
-        use std::sync::atomic::{AtomicU32, Ordering};
-        use std::sync::Mutex;
-        use windows::Win32::Foundation::{HWND, LPARAM, LRESULT, WPARAM};
-        use windows::Win32::UI::WindowsAndMessaging::{
-            CallWindowProcW, DefWindowProcW, EnumChildWindows, GetClassNameW, GetWindowLongPtrW,
-            GetWindowLongW, SetWindowLongPtrW, SetWindowLongW, SetWindowPos, GWLP_WNDPROC,
-            GWL_EXSTYLE, GWL_STYLE, SWP_FRAMECHANGED, SWP_NOACTIVATE, SWP_NOMOVE, SWP_NOSIZE,
-            SWP_NOZORDER, WNDPROC, WS_BORDER, WS_CAPTION, WS_CHILD, WS_DLGFRAME, WS_EX_CLIENTEDGE,
-            WS_EX_CONTEXTHELP, WS_EX_DLGMODALFRAME, WS_EX_STATICEDGE, WS_EX_WINDOWEDGE, WS_POPUP,
-            WS_THICKFRAME,
-        };
-
-        // WM_NCHITTEST message constant
-        const WM_NCHITTEST: u32 = 0x0084;
-        // HTCLIENT - indicates the client area (no dragging)
-        const HTCLIENT: isize = 1;
-
-        // Store original window procedures for subclassed windows
-        // Using a static HashMap protected by Mutex for thread safety
-        static ORIGINAL_WNDPROCS: Mutex<Option<HashMap<isize, isize>>> = Mutex::new(None);
-
-        // Initialize the HashMap if needed
-        {
-            // Use ok() to avoid panic if mutex is poisoned during shutdown
-            if let Ok(mut guard) = ORIGINAL_WNDPROCS.lock() {
-                if guard.is_none() {
-                    *guard = Some(HashMap::new());
-                }
-            }
-        }
-
-        // Custom window procedure that intercepts WM_NCHITTEST
-        unsafe extern "system" fn subclass_wndproc(
-            hwnd: HWND,
-            msg: u32,
-            wparam: WPARAM,
-            lparam: LPARAM,
-        ) -> LRESULT {
-            // Intercept WM_NCHITTEST to prevent dragging
-            if msg == WM_NCHITTEST {
-                // Always return HTCLIENT to indicate we're in the client area
-                // This prevents any part of the window from being treated as a drag handle
-                return LRESULT(HTCLIENT);
-            }
-
-            // Get the original window procedure - use ok() to avoid panic during shutdown
-            let original_wndproc = ORIGINAL_WNDPROCS.lock().ok().and_then(|guard| {
-                guard
-                    .as_ref()
-                    .and_then(|map| map.get(&(hwnd.0 as isize)).copied())
-            });
-
-            if let Some(original) = original_wndproc {
-                // Call the original window procedure for all other messages
-                let wndproc: WNDPROC = std::mem::transmute(original);
-                CallWindowProcW(wndproc, hwnd, msg, wparam, lparam)
-            } else {
-                // Fallback to DefWindowProc if original not found
-                DefWindowProcW(hwnd, msg, wparam, lparam)
-            }
-        }
-
-        // Counter for fixed windows
-        static FIXED_COUNT: AtomicU32 = AtomicU32::new(0);
-        static TOTAL_COUNT: AtomicU32 = AtomicU32::new(0);
-        static SUBCLASSED_COUNT: AtomicU32 = AtomicU32::new(0);
-        FIXED_COUNT.store(0, Ordering::SeqCst);
-        TOTAL_COUNT.store(0, Ordering::SeqCst);
-        SUBCLASSED_COUNT.store(0, Ordering::SeqCst);
-
-        // Callback function for EnumChildWindows
-        // Returns TRUE (non-zero) to continue enumeration, FALSE (0) to stop
-        unsafe extern "system" fn enum_child_proc(
-            child_hwnd: HWND,
-            _lparam: LPARAM,
-        ) -> windows::core::BOOL {
-            TOTAL_COUNT.fetch_add(1, Ordering::SeqCst);
-
-            // Get window class name for logging
-            let mut class_name_buf = [0u16; 256];
-            let class_len = GetClassNameW(child_hwnd, &mut class_name_buf);
-            let class_name = if class_len > 0 {
-                String::from_utf16_lossy(&class_name_buf[..class_len as usize])
-            } else {
-                String::from("<unknown>")
-            };
-
-            // Get current styles
-            let style = GetWindowLongW(child_hwnd, GWL_STYLE);
-            let ex_style = GetWindowLongW(child_hwnd, GWL_EXSTYLE);
-
-            // Check if this window has problematic styles
-            let has_popup = (style & WS_POPUP.0 as i32) != 0;
-            let has_caption = (style & WS_CAPTION.0 as i32) != 0;
-            let has_thickframe = (style & WS_THICKFRAME.0 as i32) != 0;
-            let is_child = (style & WS_CHILD.0 as i32) != 0;
-
-            tracing::info!(
-                "[fix_webview2_child_windows] Checking child HWND 0x{:X} class='{}' style=0x{:08X} (popup={}, caption={}, thickframe={}, is_child={})",
-                child_hwnd.0 as isize,
-                class_name,
-                style,
-                has_popup,
-                has_caption,
-                has_thickframe,
-                is_child
-            );
-
-            // Subclass Chrome_WidgetWin_0 and Chrome_WidgetWin_1 to intercept WM_NCHITTEST
-            // These are the windows that handle mouse input and may cause dragging
-            if class_name == "Chrome_WidgetWin_0" || class_name == "Chrome_WidgetWin_1" {
-                // Check if already subclassed - use ok() to avoid panic during shutdown
-                let already_subclassed = ORIGINAL_WNDPROCS
-                    .lock()
-                    .ok()
-                    .and_then(|guard| {
-                        guard
-                            .as_ref()
-                            .map(|map| map.contains_key(&(child_hwnd.0 as isize)))
-                    })
-                    .unwrap_or(false);
-
-                if !already_subclassed {
-                    // Get the current window procedure
-                    let original_wndproc = GetWindowLongPtrW(child_hwnd, GWLP_WNDPROC);
-                    if original_wndproc != 0 {
-                        // Store the original window procedure - use ok() to avoid panic
-                        if let Ok(mut guard) = ORIGINAL_WNDPROCS.lock() {
-                            if let Some(map) = guard.as_mut() {
-                                map.insert(child_hwnd.0 as isize, original_wndproc);
-                            }
-                        }
-
-                        // Set our custom window procedure
-                        SetWindowLongPtrW(
-                            child_hwnd,
-                            GWLP_WNDPROC,
-                            subclass_wndproc as *const () as usize as isize,
-                        );
-
-                        SUBCLASSED_COUNT.fetch_add(1, Ordering::SeqCst);
-
-                        tracing::info!(
-                            "[OK] [fix_webview2_child_windows] Subclassed HWND 0x{:X} class='{}' to intercept WM_NCHITTEST",
-                            child_hwnd.0 as isize,
-                            class_name
-                        );
-                    }
-                }
-            }
-
-            // Set dark background brush on every child window to prevent white
-            // edges from appearing when the window is partially painted.
-            auroraview_core::builder::set_window_class_dark_background(child_hwnd.0 as isize);
-
-            // Only fix windows that aren't already proper child windows
-            if has_popup || has_caption || has_thickframe || !is_child {
-                // Remove problematic styles and ensure WS_CHILD
-                let new_style = (style
-                    & !(WS_POPUP.0 as i32)
-                    & !(WS_CAPTION.0 as i32)
-                    & !(WS_THICKFRAME.0 as i32)
-                    & !(WS_BORDER.0 as i32)
-                    & !(WS_DLGFRAME.0 as i32))
-                    | (WS_CHILD.0 as i32);
-
-                // Remove extended styles that can cause white borders (match window_style.rs)
-                let new_ex_style = ex_style
-                    & !(WS_EX_STATICEDGE.0 as i32)
-                    & !(WS_EX_CLIENTEDGE.0 as i32)
-                    & !(WS_EX_WINDOWEDGE.0 as i32)
-                    & !(WS_EX_DLGMODALFRAME.0 as i32)
-                    & !(WS_EX_CONTEXTHELP.0 as i32);
-
-                if new_style != style || new_ex_style != ex_style {
-                    SetWindowLongW(child_hwnd, GWL_STYLE, new_style);
-                    SetWindowLongW(child_hwnd, GWL_EXSTYLE, new_ex_style);
-
-                    // Subclass to intercept WM_NCCALCSIZE and force zero NC area
-                    auroraview_core::builder::subclass_for_zero_nc_area(child_hwnd.0 as isize);
-
-                    // Apply changes with SWP_FRAMECHANGED
-                    let _ = SetWindowPos(
-                        child_hwnd,
-                        None,
-                        0,
-                        0,
-                        0,
-                        0,
-                        SWP_NOMOVE | SWP_NOSIZE | SWP_NOZORDER | SWP_NOACTIVATE | SWP_FRAMECHANGED,
-                    );
-
-                    FIXED_COUNT.fetch_add(1, Ordering::SeqCst);
-
-                    tracing::info!(
-                        "[OK] [fix_webview2_child_windows] Fixed child HWND 0x{:X} class='{}' (style 0x{:08X} -> 0x{:08X})",
-                        child_hwnd.0 as isize,
-                        class_name,
-                        style,
-                        new_style
-                    );
-                }
-            }
-
-            // Continue enumeration (TRUE = 1)
-            windows::core::BOOL::from(true)
-        }
-
-        unsafe {
-            let hwnd_win = HWND(hwnd as *mut c_void);
-            let _ = EnumChildWindows(Some(hwnd_win), Some(enum_child_proc), LPARAM(0));
-            let total = TOTAL_COUNT.load(Ordering::SeqCst);
-            let fixed = FIXED_COUNT.load(Ordering::SeqCst);
-            let subclassed = SUBCLASSED_COUNT.load(Ordering::SeqCst);
-            tracing::info!(
-                "[OK] [NativeBackend] Fixed WebView2 child windows for HWND 0x{:X} (total={}, fixed={}, subclassed={})",
-                hwnd,
-                total,
-                fixed,
-                subclassed
-            );
-        }
-    }
-
     /// Process messages for DCC integration mode
     ///
     /// This method should be called periodically from a Qt timer to process
@@ -1007,7 +765,7 @@ impl NativeBackend {
             }
 
             // Fix WebView2 child windows: strip border/edge styles to remove white borders
-            Self::fix_webview2_child_windows(hwnd_value);
+            auroraview_core::builder::fix_webview2_child_windows(hwnd_value);
 
             // For Child (Qt embedding): apply anti-white-border measures
             if config.embed_mode == EmbedMode::Child {
